+++ exo5.py
@@ -2,10 +2,10 @@ longueur = input("Donné la longueur du rectangle")
 largeur = input("Donné la largeur du rectangle")
 deux = int(2)
 
-peri = deux*(int(longueur)+int(largeur))
+peri = deux*(float(longueur)+float(largeur))
 
 
-aire = int(longueur) * int(largeur)
+aire = float(longueur) * float(largeur)
 
 print(peri)
